refactor(test): rename `general-utils` to `general-util` and update all references in tests
