+++ typescript/test/e2e/get-transaction-record.e2e.test.ts
@@ -8,7 +8,7 @@ import {
   getCustomClient,
 } from '../utils';
 import { Client, TransactionId, PrivateKey } from '@hashgraph/sdk';
-import { extractObservationFromLangchainResponse, wait } from '../utils/general-utils';
+import { extractObservationFromLangchainResponse, wait } from '../utils/general-util';
 import Long from 'long';
 
 describe('Get Transaction Record E2E Tests', () => {
