+++ typescript/test/integration/hedera/create-topic.integration.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeAll, afterAll } from 'vitest';
 import { Client } from '@hashgraph/sdk';
 import createTopicTool from '@/plugins/core-consensus-plugin/tools/consensus/create-topic';
 import { Context, AgentMode } from '@/shared/configuration';
-import { getClientForTests, HederaOperationsWrapper } from '../../utils';
+import { getOperatorClientForTests, HederaOperationsWrapper } from '../../utils';
 import { z } from 'zod';
 import { createTopicParameters } from '@/shared/parameter-schemas/consensus.zod';
 
@@ -12,7 +12,7 @@ describe('Create Topic Integration Tests', () => {
   let hederaOperationsWrapper: HederaOperationsWrapper;
 
   beforeAll(async () => {
-    client = getClientForTests();
+    client = getOperatorClientForTests();
     hederaOperationsWrapper = new HederaOperationsWrapper(client);
 
     context = {
